Add failing infrastructure plugin for runtime tests
