+++ agents.log
@@ -1,3 +1,4 @@
+AGENT NOTE - 2025-07-15: Implemented failing infrastructure plugin for tests
 <<<<<<< HEAD
 AGENT NOTE - 2025-07-15: Updated memory._execute to handle paramstyle fallback
 <<<<<<< HEAD
