+++ tests/infrastructure/__init__.py
@@ -1,0 +1,8 @@
+"""Infrastructure helpers used in tests."""
+
+from .failing_plugins import FailingDuckDBInfrastructure, FailingPostgresInfrastructure
+
+__all__ = [
+    "FailingDuckDBInfrastructure",
+    "FailingPostgresInfrastructure",
+]
