+++ tests/infrastructure/failing_plugins.py
@@ -1,0 +1,36 @@
+from __future__ import annotations
+
+from contextlib import asynccontextmanager
+
+from entity.infrastructure import DuckDBInfrastructure, PostgresInfrastructure
+
+
+class FailingDuckDBInfrastructure(DuckDBInfrastructure):
+    """DuckDB plugin that always fails during runtime validation."""
+
+    name = "failing_duckdb"
+
+    @asynccontextmanager
+    async def connection(self):  # type: ignore[override]
+        class BadConn:
+            def execute(self, _q):
+                raise RuntimeError("boom")
+
+            def close(self):
+                pass
+
+        yield BadConn()
+
+
+class FailingPostgresInfrastructure(PostgresInfrastructure):
+    """Postgres plugin that always fails during runtime validation."""
+
+    name = "failing_postgres"
+
+    @asynccontextmanager
+    async def connection(self):  # type: ignore[override]
+        class BadConn:
+            async def execute(self, *args, **kwargs):
+                raise RuntimeError("bad")
+
+        yield BadConn()
