+++ tests/test_infrastructure_validate_runtime.py
@@ -1,26 +1,25 @@
-from contextlib import asynccontextmanager
 import pytest
-from entity.infrastructure import DuckDBInfrastructure, PostgresInfrastructure
+
 from entity.core.circuit_breaker import CircuitBreaker
+from entity.core.resources.container import ResourceContainer
+from tests.infrastructure import (
+    FailingDuckDBInfrastructure,
+    FailingPostgresInfrastructure,
+)
 
 
 @pytest.mark.asyncio
-async def test_duckdb_runtime_breaker_opens(monkeypatch):
-    db = DuckDBInfrastructure({"failure_threshold": 2})
+async def test_duckdb_runtime_breaker_opens():
+    container = ResourceContainer()
+    container.register(
+        "duckdb",
+        FailingDuckDBInfrastructure,
+        {"failure_threshold": 2},
+        layer=1,
+    )
+    await container.build_all()
+    db = container.get("duckdb")
     breaker = CircuitBreaker(failure_threshold=2)
-
-    class BadConn:
-        def execute(self, _q):
-            raise RuntimeError("boom")
-
-        def close(self):
-            pass
-
-    @asynccontextmanager
-    async def bad_connection(self):
-        yield BadConn()
-
-    monkeypatch.setattr(DuckDBInfrastructure, "connection", bad_connection)
 
     res1 = await db.validate_runtime(breaker)
     assert not res1.success
@@ -33,20 +32,21 @@ async def test_duckdb_runtime_breaker_opens(monkeypatch):
     assert not res3.success
     assert "circuit breaker open" in res3.message.lower()
 
+    await container.shutdown_all()
+
 
 @pytest.mark.asyncio
 async def test_postgres_runtime_breaker_opens():
-    pg = PostgresInfrastructure({"failure_threshold": 2})
+    container = ResourceContainer()
+    container.register(
+        "postgres",
+        FailingPostgresInfrastructure,
+        {"failure_threshold": 2},
+        layer=1,
+    )
+    await container.build_all()
+    pg = container.get("postgres")
     breaker = CircuitBreaker(failure_threshold=2)
-
-    class BadConn:
-        async def execute(self, *args, **kwargs):
-            raise RuntimeError("bad")
-
-    async def acquire():
-        return BadConn()
-
-    pg._pool.acquire = acquire
 
     res1 = await pg.validate_runtime(breaker)
     assert not res1.success
@@ -58,3 +58,5 @@ async def test_postgres_runtime_breaker_opens():
     res3 = await pg.validate_runtime(breaker)
     assert not res3.success
     assert "circuit breaker open" in res3.message.lower()
+
+    await container.shutdown_all()
